Structure & organise file
